# perf: faster escape html & unescape html