--- a/test/escape_html.spec.ts
+++ b/test/escape_html.spec.ts
@@ -4,7 +4,7 @@ chai.should();
 
 describe('escapeHTML', () => {
   it('default', () => {
-    escapeHTML('<p class="foo">Hello "world".</p>').should.eql('&lt;p class&#x3D;&quot;foo&quot;&gt;Hello &quot;world&quot;.&lt;&#x2F;p&gt;');
+    escapeHTML('<p class="foo">Hello `world`.</p>').should.eql('&lt;p class&#x3D;&quot;foo&quot;&gt;Hello &#96;world&#96;.&lt;&#x2F;p&gt;');
   });
 
   it('str must be a string', () => {
@@ -12,7 +12,7 @@ describe('escapeHTML', () => {
   });
 
   it('avoid double escape', () => {
-    escapeHTML('&lt;foo>bar</foo&gt;').should.eql('&lt;foo&gt;bar&lt;&#x2F;foo&gt;');
+    escapeHTML('&lt;foo>bar</foo&gt;&#x2F;').should.eql('&lt;foo&gt;bar&lt;&#x2F;foo&gt;&#x2F;');
   });
 
   it('avoid double escape https://github.com/hexojs/hexo/issues/4946', () => {