--- a/test/escape_html.spec.ts
+++ b/test/escape_html.spec.ts
@@ -12,7 +12,7 @@ describe('escapeHTML', () => {
   });
 
   it('avoid double escape', () => {
-    escapeHTML('&lt;foo>bar</foo&gt;&#x2F;').should.eql('&lt;foo&gt;bar&lt;&#x2F;foo&gt;&#x2F;');
+    escapeHTML('&lt;foo>bar</foo&gt;&#x2F;|&6>').should.eql('&lt;foo&gt;bar&lt;&#x2F;foo&gt;&#x2F;|&6&gt;');
   });
 
   it('avoid double escape https://github.com/hexojs/hexo/issues/4946', () => {