--- a/test/escape_html.spec.ts
+++ b/test/escape_html.spec.ts
@@ -12,10 +12,14 @@ describe('escapeHTML', () => {
   });
 
   it('avoid double escape', () => {
-    escapeHTML('&lt;foo>bar</foo&gt;&#x2F;|&6>').should.eql('&lt;foo&gt;bar&lt;&#x2F;foo&gt;&#x2F;|&6&gt;');
+    escapeHTML('&lt;foo>bar</foo&gt;&#x2F;|&6>').should.eql('&lt;foo&gt;bar&lt;&#x2F;foo&gt;&#x2F;|&amp;6&gt;');
   });
 
   it('avoid double escape https://github.com/hexojs/hexo/issues/4946', () => {
     escapeHTML('&emsp;&nbsp;&ensp;').should.eql('&emsp;&nbsp;&ensp;');
   });
+
+  it('proper escape', () => {
+    escapeHTML('&0').should.eql('&amp;0');
+  })
 });